refactor: 코드 수정
(cherry picked from commit e8c760093f45a4da1e277730abc92c4490e26215)
