+++ src/main/java/domain/ladder/LadderWidth.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 public class Width {
